+++ fathom-rest/src/main/java/fathom/rest/Context.java
@@ -17,13 +17,18 @@
 package fathom.rest;
 
 import ro.pippo.core.Application;
+import ro.pippo.core.HttpConstants;
 import ro.pippo.core.Request;
 import ro.pippo.core.Response;
 import ro.pippo.core.route.DefaultRouteContext;
 import ro.pippo.core.route.RouteHandler;
 import ro.pippo.core.route.RouteMatch;
+import ro.pippo.core.util.StringUtils;
 
+import java.util.Collections;
+import java.util.LinkedHashSet;
 import java.util.List;
+import java.util.Set;
 
 /**
  * @author James Moger
@@ -36,6 +41,50 @@ public class Context extends DefaultRouteContext {
 
     public Context(Context context, List<RouteMatch> routeMatches) {
         this(context.getApplication(), context.getRequest(), context.getResponse(), routeMatches);
+    }
+
+    public Set<String> getAcceptTypes() {
+        Set<String> types = new LinkedHashSet<>();
+        types.addAll(getContentTypes(getRequest().getAcceptType()));
+        types.addAll(getContentTypes(getRequest().getHttpServletRequest().getHeader(HttpConstants.Header.ACCEPT)));
+        return types;
+    }
+
+    public Set<String> getContentTypes() {
+        Set<String> types = new LinkedHashSet<>();
+        types.addAll(getContentTypes(getRequest().getContentType()));
+        types.addAll(getContentTypes(getRequest().getHttpServletRequest().getContentType()));
+        return types;
+    }
+
+    /**
+     * Cleans a complex content-type or accept header value by removing the
+     * quality scores.
+     * <p/>
+     * <pre>
+     * text/html,application/xhtml+xml,application/xml;q=0.9,image/webp
+     * </pre>
+     *
+     * @param contentType
+     * @return the sanitized set of content-types
+     */
+    protected Set<String> getContentTypes(String contentType) {
+        if (StringUtils.isNullOrEmpty(contentType)) {
+            return Collections.emptySet();
+        }
+
+        Set<String> set = new LinkedHashSet<>();
+        String[] types = contentType.split(",");
+        for (String type : types) {
+            if (type.contains(";")) {
+                // drop ;q=0.8 quality scores
+                type = type.substring(0, type.indexOf(';'));
+            }
+
+            set.add(type);
+        }
+
+        return set;
     }
 
 }